+++ src/SFA.DAS.Roatp.Database/Tables/NationalAchievementRateOverall.sql
@@ -1,9 +1,9 @@
 ﻿CREATE TABLE [dbo].[NationalAchievementRateOverall]
 (
 	[Id] BIGINT IDENTITY (1,1),
-	[Age] SMALLINT NOT NULL DEFAULT 0,
+	[Age] INT NOT NULL DEFAULT 0,
 	[SectorSubjectArea] VARCHAR(1000) NOT NULL,
-	[ApprenticeshipLevel] SMALLINT NOT NULL DEFAULT 0,
+	[ApprenticeshipLevel] INT NOT NULL DEFAULT 0,
 	[OverallCohort] INT NULL,
 	[OverallAchievementRate] decimal(10,4) NULL,
 	CONSTRAINT PK_NationalAchievementRateOverall PRIMARY KEY (Id)
